Remove the need to pass the ID index instance around
Make manual_ids() a self-contained lump, as there should never be a need
to have more than one index -- making the passing of the instance to the
output routines a nuisance.

Error checking of parse_link() could be improved, as there is currently
no feedback to the caller of whether things worked or not.