Prevent blocks from starting block collections
Block collections (currently <li>) can't start with <ul>, <ol>, <code>,
or <table>.

This is because in text mode, these objects would wish to push a new
line on to a line which might already exist: eg. a list would have
created a line containing the bullet and be requesting that the first
object goes into one of the columns.

In general, such items will need to start with a paragraph anyway, so
this shouldn't be an imposition.