Include Stylesheet links in HTML documents
Provide a new <stylesheet> tag in the HTML resources block to specify
a filename for a stypesheet which will apply to all HTML documents from
that point down the tree.

The filename is relative to the point that it is specified, and will be
adjusted for any file nesting lower down the tree.