Add Callout support
Add a <callout> tag, which will write a callout box in all three modes.
Currently, StrongHelp mode has placeholder support.

A callout can contain a <title>, along with paragraphs, lists and
code blocks. If no title is specified, a default title is used based on
the callout type. Types are:

 - Attention
 - Caution
 - Danger
 - Error
 - Hint
 - Important
 - Note
 - See Also
 - Tip
 - Warning