Implement command line defines
Add a -define "name=value" parameter to the command line options, which
can be used to pass in externally-defined text for inclusion in the
manual. This can then be accessed with a <define name="name"/> tag
within the manual body.

Currently the text is a plain UTF-8 string, with no attempt to parse
entities contained within it.