Fix handling of &msep; entity
Add a special case for our home-grown &msep; entity, so that we don't
get a crash when processing manuals containing it.

The entity currently maps to an endash, but this should be fixed so that
it maps to a right-pointing arrow when supported (StrongHelp, HTML).
Arrow entities are currently disabled due to duplicate names, so this
needs to be fixed at the same time.