Don't assume that a codepoint will fit in a char
If we're in UTF-8 output mode, we can't assume that we can safely write
a codepoint to a char, so fall back to writing ? instead.

In practice, the will never be executed as encoding_write_unicode_char()
will not call encoding_find_mapped_character() if there isn't an output
encoding specified. But it's better not to leave the trap for the unwary
in the future.